requirements.txt
anyio==4.11.0
av==16.0.1
certifi==2025.11.12
click==8.3.0
coloredlogs==15.0.1
ctranslate2==4.6.1
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-genai>=0.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.3.4
# nvidia-cublas-cu12==12.9.1.4  # torch gestiona esto automáticamente
# nvidia-cudnn-cu12==9.16.0.29  # torch gestiona esto automáticamente
onnxruntime==1.23.2
packaging==25.0
protobuf==6.33.1
python-dotenv>=1.0.0
PyYAML==6.0.3
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tenacity>=8.2.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
wheel==0.45.1
toml
python-json-logger
pydantic-settings
sounddevice
silero-vad
pytest
pytest-asyncio
pytest-mock
PySide6
